refactor(clientDto): changed clientDto location
